Refactor navigator assembly responsibilities
