+++ app/service/navigator_runtime/__init__.py
@@ -3,6 +3,12 @@ from __future__ import annotations
 
 from .assembly import build_runtime_from_dependencies
 from .entrypoints import assemble_navigator
+from .presentation import (
+    NavigatorRuntimeProvider,
+    RuntimeAssemblyConfiguration,
+    RuntimeAssemblyEntrypoint,
+    default_configuration,
+)
 from .runtime_factory import (
     NavigatorRuntimeAssembly,
     RuntimePlannerDependencies,
@@ -17,12 +23,16 @@ from .usecases import NavigatorUseCases
 __all__ = [
     "NavigatorRuntime",
     "NavigatorRuntimeAssembly",
+    "NavigatorRuntimeProvider",
     "NavigatorUseCases",
     "RuntimePlannerDependencies",
+    "RuntimeAssemblyConfiguration",
+    "RuntimeAssemblyEntrypoint",
     "assemble_navigator",
     "build_navigator_runtime",
     "build_runtime_collaborators",
     "build_runtime_contract_selection",
     "build_runtime_from_dependencies",
+    "default_configuration",
     "create_runtime_plan_request",
 ]
